Travis: run the tests against PHP 8/nightly
Now the tests are fully compatible with PHPUnit 9, the test suite can be run on PHP 8.

As not all dependencies of PHPUnit have tagged a new version which allows for PHP 8 yet, we do need to `ignore-platform-reqs` for the time being.

Includes adding `travis_retry` to `composer install`-like commands to get round builds stalling on the connection with Packagist, especially on older PHP versions.